# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedRise Medical Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
        - name: staffId
          in: query
          required: false
          schema:
            type: integer
        - name: month
          in: query
          required: false
          description: "Format: YYYY-MM"
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: recordAttendance
      tags: [attendance]
      summary: Record or update attendance for a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "200":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "404":
          description: Not found
    delete:
      operationId: deleteAttendance
      tags: [attendance]
      summary: Delete an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /attendance/stats/summary:
    get:
      operationId: getAttendanceStats
      tags: [attendance]
      summary: Attendance statistics
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceStats"
...
  /billing:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: List invoices
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [billing]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/{id}:
    get:
      operationId: getInvoice
      tags: [billing]
      summary: Get a single invoice with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    patch:
      operationId: updateInvoice
      tags: [billing]
      summary: Update invoice status or payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    delete:
      operationId: deleteInvoice
      tags: [billing]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    AttendanceRecord:
      type: object
      required: [id, staffId, date, shift, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        staffRole:
          type: ["string", "null"]
        staffTitle:
          type: ["string", "null"]
        date:
          type: string
        shift:
          type: string
        status:
          type: string
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        recordedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    AttendanceInput:
      type: object
      required: [staffId, date, status]
      properties:
        staffId:
          type: integer
        date:
          type: string
        shift:
          type: string
          enum: [day, night, morning, afternoon]
        status:
          type: string
          enum: [present, absent, late, leave, off]
        checkIn:
          type: string
        checkOut:
          type: string
        notes:
          type: string
    AttendanceUpdateInput:
      type: object
      properties:
        shift:
          type: string
          enum: [day, night, morning, afternoon]
        status:
          type: string
          enum: [present, absent, late, leave, off]
        checkIn:
          type: string
        checkOut:
          type: string
        notes:
          type: string
    AttendanceStats:
      type: object
      required: [totalDays, totalPresent, totalAbsent, totalLate, totalLeave, totalOff, staffSummary]
      properties:
        totalDays:
          type: integer
        totalPresent:
          type: integer
        totalAbsent:
          type: integer
        totalLate:
          type: integer
        totalLeave:
          type: integer
        totalOff:
          type: integer
        staffSummary:
          type: array
          items:
            type: object
            required: [staffId, staffName, present, absent, late, leave, off]
            properties:
              staffId:
                type: integer
              staffName:
                type: string
              staffRole:
                type: string
              present:
                type: integer
              absent:
                type: integer
              late:
                type: integer
              leave:
                type: integer
              off:
                type: integer
...
    InvoiceItem:
      type: object
      required: [id, invoiceId, description, quantity, unitPrice, amount]
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: string
        amount:
          type: string
    Invoice:
      type: object
      required: [id, patientId, invoiceNumber, status, totalAmount, paidAmount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        consultationId:
          type: ["integer", "null"]
        invoiceNumber:
          type: string
        status:
          type: string
          enum: [unpaid, partial, paid, cancelled]
        totalAmount:
          type: string
        paidAmount:
          type: string
        paymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
        updatedAt:
          type: string
    InvoiceInput:
      type: object
      required: [patientId, items]
      properties:
        patientId:
          type: integer
        consultationId:
          type: integer
        notes:
          type: string
        items:
          type: array
          items:
            type: object
            required: [description, quantity, unitPrice]
            properties:
              description:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
    InvoiceUpdateInput:
      type: object
      properties:
        status:
          type: string
          enum: [unpaid, partial, paid, cancelled]
        paidAmount:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
...